fix: move runtime hints to storage module.
