+++ spring-cloud-gcp-storage/src/main/resources/META-INF/spring/aot.factories
@@ -1,0 +1,2 @@
+org.springframework.aot.hint.RuntimeHintsRegistrar=\
+com.google.cloud.spring.storage.integration.aot.StorageIntegrationRuntimeHint
